requirements.txt
geopandas
shapely
pandas
pyogrio
mcp
python-dotenv
requests
uvicorn
httpx
pytest
anthropic
